git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@379 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/FILTERING/SMOOTHING/SavitzkyGolaySVDFilter.h
@@ -183,7 +183,7 @@ namespace OpenMS
       template < typename InputPeakIterator, typename OutputPeakContainer  >
       void filter(InputPeakIterator first, InputPeakIterator last, OutputPeakContainer& smoothed_data_container) throw (Exception::InvalidSize)
       {
-        if (distance(first,last) <= frame_size_)
+        if (distance(first,last) <= (int)frame_size_)
         {
           throw Exception::InvalidSize(__FILE__,__LINE__,__PRETTY_FUNCTION__,distance(first,last));
         }
